shop_backend/requirements.txt
djangorestframework
Django~=2.2.4
django_cors_headers